+++ CHANGES.md
@@ -12,6 +12,7 @@ Changes in 4.6
 - Add Comment column to Activities list in Activities.php
 - Format PHP code, program wide
 - Use StudentUsersInfo \_make\*Input functions in Schools.php
+- Add \_makeFilesInput() function & $options_RET parameter to \_makeAutoSelectInput() in StudentsUsersInfo.fnc.php
 
 Changes in 4.5.2
 ----------------